feat!: Move PeerDID outside of Domain.
BREAKING CHANGE: If you have imported Domain.PeerDID in your typescript application, you will need to re-import from PeerDID directly
