+++ src/index.ts
@@ -16,7 +16,7 @@ export * from "./prism-agent/mediator/PlutoMediatorStore";
 export * from "./mercury/didcomm/Wrapper";
 export * from "./prism-agent/helpers/ApiImpl";
 export { ListenerKey } from "./prism-agent/types";
-
+export * from './peer-did/PeerDID';
 export type {
   MediatorHandler,
   ConnectionsManager as ConnectionsManagerInterface,
