+++ src/peer-did/PeerDID.ts
@@ -1,11 +1,42 @@
-import { DID } from "../domain/models/DID";
-import { CastorError } from "../domain/models/Errors";
+import { CastorError, type DID, type KeyCurve } from "../domain";
+
+export namespace PeerDID {
+  // Q: why is this a custom shape instead of a Domain.PrivateKey?
+  export interface PrivateKey {
+    /**
+     * Instance of a KeyCurve
+     *
+     * @type {KeyCurve}
+     */
+    keyCurve: KeyCurve;
+    /**
+     * Value as Uint8Array, buffer like
+     *
+     * @type {Uint8Array}
+     */
+    value: Uint8Array;
+  }
+}
 
 export interface PeerDIDEncoded {
   t: string;
   s: string;
   r: string[];
   a: string[];
+}
+
+export class PeerDID {
+  constructor(
+    public readonly did: DID,
+    public readonly privateKeys: PeerDID.PrivateKey[] = []
+  ) {
+    const regex = /(([01](z)([1-9a-km-zA-HJ-NP-Z]{46,47}))|(2((\.[AEVID](z)([1-9a-km-zA-HJ-NP-Z]{46,47}))+(\.(S)[0-9a-zA-Z=]*)?)))$/;
+    const isValid = did.schema === "did" && did.method === "peer" && regex.test(did.methodId);
+
+    if (isValid === false) {
+      throw new CastorError.InvalidPeerDIDError();
+    }
+  }
 }
 
 /**
@@ -60,24 +91,5 @@ export class PeerDIDService {
       encoded.r,
       encoded.a
     );
-  }
-}
-
-export class PeerDID {
-  readonly did: DID;
-
-  constructor(did: DID) {
-    const regex =
-      /(([01](z)([1-9a-km-zA-HJ-NP-Z]{46,47}))|(2((\.[AEVID](z)([1-9a-km-zA-HJ-NP-Z]{46,47}))+(\.(S)[0-9a-zA-Z=]*)?)))$/;
-    if (
-      !(
-        did.schema === "did" &&
-        did.method === "peer" &&
-        regex.test(did.methodId)
-      )
-    ) {
-      throw new CastorError.InvalidPeerDIDError();
-    }
-    this.did = did;
   }
 }
